Feature - Dockerization of the lambda dependencies
This is an "untested" dockerization of the imagemagick binary dependencies. It also includes a WIP shell script for launching the dockerization as well as actually copying over the build dependencies to the local file system. Note: might be more "JavaScript-y" if I convert the bash script into a pure JS file/function.