Feature - ImageMagick binary switching refactor
WIP: Major refactor to enable local testing and switching of which imagemagick binary to use. This refactor breaks up the monolithic main.js file into a series of dependencies, and calls the actual C imagemagick process using `child_process` instead of wrapping it in the `gm` library, which is awful.

Closes #16