Bugfix - Fixed S3 uploads of modified images.
Various bugs were present causing issues with finding and using the images. Among them, the async implemention of S3 was terrible and had to be re-written, the pointer to the file location was incorrect and causing a loop to occur becuase all S3 uploads were happening to the same /original/ location, improper return values were used, and the install script was doing unnecessary work.

Closes #25