Feature - Response Object
This PR is a refactor of the main lambda function to break up the individual logic pieces into asyncronous control flow methods. It creates two imports, resize and format, which are where the imagemagick logic takes place, and which return async functions. This also creates the conversion options in JSON files and modifies the eslint rules to have node as a valid target.

Closes #7