Bugfix - body.params error
This is most likely the fix for the intermittent errors we have seen in Magine. When the JSON files were getting uploaded to S3, it was triggering the lambda as well, and then trying to perform imagemagick stuff on it, resulting in failures that were only caught when it tried to re-upload the files to the new S3 location. This is a dirty fix, but it allows for the execution to halt ASAP, as opposed to handling it further down the line when it has already loaded in the imagemagick libs and started to allocate heap resources.