Bugfix - Serve most optimized images first
This changes the order of the conversion params so that the msot optimized and future facing ones are first in the list, and hopefully, first in the `types` array returned by this lambda. This effects both the frontend and the cms, as they are relying on the types array. There shouldn't be any changes necessary for the intended effect in those repos though.